Plot celda Heatmap

Render a stylable heatmap of count data based on celda clustering results.

celdaHeatmap(
  sce,
  useAssay = "counts",
  altExpName = "featureSubset",
  featureIx = NULL,
  nfeatures = 25,
  ...
)

# S4 method for SingleCellExperiment
celdaHeatmap(
  sce,
  useAssay = "counts",
  altExpName = "featureSubset",
  featureIx = NULL,
  nfeatures = 25,
  ...
)

Arguments

sce

A SingleCellExperiment object returned by celda_C, celda_G, or celda_CG.

useAssay

A string specifying which assay slot to use. Default "counts".

altExpName

The name for the altExp slot to use. Default "featureSubset".

featureIx

Integer vector. Select features for display in heatmap. If NULL, no subsetting will be performed. Default NULL. Only used for sce containing celda_C model result returned by celda_C.

nfeatures

Integer. Maximum number of features to select for each gene module. Default 25. Only used for sce containing celda_CG or celda_G model results returned by celda_CG or celda_G.

...

Additional parameters passed to plotHeatmap.

Value

list A list containing dendrogram information and the heatmap grob

See also

Examples

data(sceCeldaCG)
celdaHeatmap(sceCeldaCG)
#> TableGrob (5 x 6) "layout": 9 grobs
#>   z     cells                 name                       grob
#> 1 1 (2-2,3-3)             col_tree polyline[GRID.polyline.16]
#> 2 2 (4-4,1-1)             row_tree polyline[GRID.polyline.17]
#> 3 3 (4-4,3-3)               matrix       gTree[GRID.gTree.19]
#> 4 4 (3-3,3-3)       col_annotation         rect[GRID.rect.20]
#> 5 5 (3-3,4-4) col_annotation_names         text[GRID.text.21]
#> 6 6 (4-4,2-2)       row_annotation         rect[GRID.rect.22]
#> 7 7 (5-5,2-2) row_annotation_names         text[GRID.text.23]
#> 8 8 (4-5,6-6)     annotationLegend       gTree[GRID.gTree.31]
#> 9 9 (4-5,5-5)               legend       gTree[GRID.gTree.34]