Get or set the feature module labels from a celda SingleCellExperiment object.

Return or set the feature module cluster labels determined by celda_G or celda_CG models.

celdaModules(sce, altExpName = "featureSubset")

# S4 method for SingleCellExperiment
celdaModules(sce, altExpName = "featureSubset")

celdaModules(sce, altExpName = "featureSubset") <- value

# S4 method for SingleCellExperiment
celdaModules(sce, altExpName = "featureSubset") <- value

Arguments

sce
A SingleCellExperiment object returned by celda_G, or celda_CG, with the matrix located in the useAssay assay slot. Rows represent features and columns represent cells.

altExpName
The name for the altExp slot to use. Default "featureSubset".

value
Character vector of feature module labels for replacements. Works only if x is a SingleCellExperiment object.

Value

Character vector. Contains feature module labels for each feature in x.

Examples

data(sceCeldaCG)
celdaModules(sceCeldaCG)

#>  [1] 7  5  1  8  1  4  10 9  2  4  7  10 10 5  3  10 2  10 9  5  9  3  5  7  6 
#> [26] 9  4  8  2  9  5  4  5  9  5  5  4  4  6  5  8  8  5  1  6  10 9  7  1  4 
#> [51] 3  7  9  6  9  10 7  6  8  6  9  5  4  5  9  9  10 5  3  8  7  9  6  2  3 
#> [76] 7  6  3  4  9  3  9  10 4  10 7  1  1  10 10 1  6  6  9  5  4 
#> Levels: 1 2 3 4 5 6 7 8 9 10