Check whether a count matrix was the one used in a given celda run

Ensures that the provided celda object was generated from a counts matrix with similar dimensions to the one provided.

compareCountMatrix(counts, celda.mod, error.on.mismatch = TRUE)

Arguments

counts
Integer matrix. Rows represent features and columns represent cells.

celda.mod
Celda model. Options available in `celda::available.models`.

error.on.mismatch
Logical. Whether to stop execution in the event of a count matrix mismatch. Default TRUE.

Value

TRUE if provided count matrix matches the one used in the celda run, FALSE otherwise. Error on FALSE if error.on.mismatch is TRUE.

Details

Then, compare the MD5 checksum of a provided count.matrix to the count matrix checksum on a celda_list object, to see if they're the same.

Examples

celda.sim = simulateCells("celda_CG")
celda.mod = celda_CG(celda.sim$counts, K=celda.sim$K, L=celda.sim$L,
                     nchains=1, max.iter=1)

#> --------------------------------------------------------------------

#> Starting Celda_CG: Clustering cells and genes.

#> --------------------------------------------------------------------

#> Thu Sep 06 12:56:44 2018 .. Initializing chain 1 with 'random' (seed=12345)

#> Thu Sep 06 12:56:44 2018 .... Determining if any cell clusters should be split.

#> Thu Sep 06 12:56:44 2018 .... Cluster 4 was reassigned and cluster 3 was split in two.

#> Thu Sep 06 12:56:44 2018 .... Determining if any gene clusters should be split.

#> Thu Sep 06 12:56:45 2018 .... Cluster 2 was reassigned and cluster 9 was split in two.

#> Thu Sep 06 12:56:45 2018 .... Completed iteration: 1 | logLik: -1272652.93151523

#> Thu Sep 06 12:56:45 2018 .. Finished chain 1 with seed 12345

#> --------------------------------------------------------------------

#> Completed Celda_CG. Total time: 0.5846629 secs

#> --------------------------------------------------------------------

 compareCountMatrix(celda.sim$counts, celda.mod, error.on.mismatch=FALSE)

#> [1] TRUE