Obtain the gene module of a gene of interest

This function will output the gene module of a specific gene(s) from a celda model

# S3 method for celda_G
featureModuleLookup(counts, celda.mod, feature,
  exact.match = TRUE)

Arguments

counts
Integer matrix. Rows represent features and columns represent cells. This matrix should be the same as the one used to generate `celda.mod`.

celda.mod
Model of class "celda_G" or "celda_CG".

feature
Character vector. Identify feature modules for the specified feature names.

exact.match
Logical. Whether to look for exact match of the gene name within counts matrix. Default TRUE.

Value

List. Each entry corresponds to the feature module determined for the provided features

Examples

celda.sim = simulateCells("celda_G")
celda.mod = celda_G(celda.sim$counts, L=celda.sim$L)

#> --------------------------------------------------------------------

#> Starting Celda_G: Clustering genes.

#> --------------------------------------------------------------------

#> Thu Sep 06 12:57:14 2018 .. Initializing chain 1 with 'random' (seed=12345)

#> Thu Sep 06 12:57:14 2018 .... Completed iteration: 1 | logLik: -291426.665356919

#> Thu Sep 06 12:57:14 2018 .... Completed iteration: 2 | logLik: -286595.332370062

#> Thu Sep 06 12:57:14 2018 .... Completed iteration: 3 | logLik: -286579.864858824

#> Thu Sep 06 12:57:14 2018 .... Completed iteration: 4 | logLik: -286579.864858824

#> Thu Sep 06 12:57:14 2018 .... Completed iteration: 5 | logLik: -286579.864858824

#> Thu Sep 06 12:57:14 2018 .... Completed iteration: 6 | logLik: -286579.864858824

#> Thu Sep 06 12:57:14 2018 .... Completed iteration: 7 | logLik: -286579.864858824

#> Thu Sep 06 12:57:14 2018 .... Completed iteration: 8 | logLik: -286579.864858824

#> Thu Sep 06 12:57:14 2018 .... Completed iteration: 9 | logLik: -286579.864858824

#> Thu Sep 06 12:57:14 2018 .... Determining if any gene clusters should be split.

#> Thu Sep 06 12:57:15 2018 .... No additional splitting was performed.

#> Thu Sep 06 12:57:15 2018 .... Completed iteration: 10 | logLik: -286579.864858824

#> Thu Sep 06 12:57:15 2018 .... Completed iteration: 11 | logLik: -286579.864858824

#> Thu Sep 06 12:57:15 2018 .... Completed iteration: 12 | logLik: -286579.864858824

#> Thu Sep 06 12:57:15 2018 .... Determining if any gene clusters should be split.

#> Thu Sep 06 12:57:15 2018 .... No additional splitting was performed.

#> Thu Sep 06 12:57:15 2018 .... Completed iteration: 13 | logLik: -286579.864858824

#> Thu Sep 06 12:57:15 2018 .. Finished chain 1 with seed 12345

#> Thu Sep 06 12:57:15 2018 .. Initializing chain 2 with 'random' (seed=12346)

#> Thu Sep 06 12:57:16 2018 .... Completed iteration: 1 | logLik: -287718.041765169

#> Thu Sep 06 12:57:16 2018 .... Completed iteration: 2 | logLik: -286579.864858824

#> Thu Sep 06 12:57:16 2018 .... Completed iteration: 3 | logLik: -286579.864858824

#> Thu Sep 06 12:57:16 2018 .... Completed iteration: 4 | logLik: -286579.864858824

#> Thu Sep 06 12:57:16 2018 .... Completed iteration: 5 | logLik: -286579.864858824

#> Thu Sep 06 12:57:16 2018 .... Completed iteration: 6 | logLik: -286579.864858824

#> Thu Sep 06 12:57:16 2018 .... Completed iteration: 7 | logLik: -286579.864858824

#> Thu Sep 06 12:57:16 2018 .... Completed iteration: 8 | logLik: -286579.864858824

#> Thu Sep 06 12:57:16 2018 .... Completed iteration: 9 | logLik: -286579.864858824

#> Thu Sep 06 12:57:16 2018 .... Determining if any gene clusters should be split.

#> Thu Sep 06 12:57:16 2018 .... No additional splitting was performed.

#> Thu Sep 06 12:57:16 2018 .... Completed iteration: 10 | logLik: -286579.864858824

#> Thu Sep 06 12:57:16 2018 .... Completed iteration: 11 | logLik: -286579.864858824

#> Thu Sep 06 12:57:16 2018 .... Determining if any gene clusters should be split.

#> Thu Sep 06 12:57:17 2018 .... No additional splitting was performed.

#> Thu Sep 06 12:57:17 2018 .... Completed iteration: 12 | logLik: -286579.864858824

#> Thu Sep 06 12:57:17 2018 .. Finished chain 2 with seed 12346

#> Thu Sep 06 12:57:17 2018 .. Initializing chain 3 with 'random' (seed=12347)

#> Thu Sep 06 12:57:17 2018 .... Completed iteration: 1 | logLik: -290090.898941072

#> Thu Sep 06 12:57:17 2018 .... Completed iteration: 2 | logLik: -286579.864858824

#> Thu Sep 06 12:57:17 2018 .... Completed iteration: 3 | logLik: -286579.864858824

#> Thu Sep 06 12:57:17 2018 .... Completed iteration: 4 | logLik: -286579.864858824

#> Thu Sep 06 12:57:17 2018 .... Completed iteration: 5 | logLik: -286579.864858824

#> Thu Sep 06 12:57:17 2018 .... Completed iteration: 6 | logLik: -286579.864858824

#> Thu Sep 06 12:57:17 2018 .... Completed iteration: 7 | logLik: -286579.864858824

#> Thu Sep 06 12:57:17 2018 .... Completed iteration: 8 | logLik: -286579.864858824

#> Thu Sep 06 12:57:17 2018 .... Completed iteration: 9 | logLik: -286579.864858824

#> Thu Sep 06 12:57:17 2018 .... Determining if any gene clusters should be split.

#> Thu Sep 06 12:57:17 2018 .... No additional splitting was performed.

#> Thu Sep 06 12:57:17 2018 .... Completed iteration: 10 | logLik: -286579.864858824

#> Thu Sep 06 12:57:17 2018 .... Completed iteration: 11 | logLik: -286579.864858824

#> Thu Sep 06 12:57:17 2018 .... Determining if any gene clusters should be split.

#> Thu Sep 06 12:57:18 2018 .... No additional splitting was performed.

#> Thu Sep 06 12:57:18 2018 .... Completed iteration: 12 | logLik: -286579.864858824

#> Thu Sep 06 12:57:18 2018 .. Finished chain 3 with seed 12347

#> --------------------------------------------------------------------

#> Completed Celda_G. Total time: 4.055695 secs

#> --------------------------------------------------------------------

module = featureModuleLookup(celda.sim$counts, celda.mod, c("Gene_1", "Gene_XXX"))