Gene set enrichment

Identify and return significantly-enriched terms for each gene module in a Celda object or a SingleCellExperiment object. Performs gene set enrichment analysis for Celda identified modules using the enrichr.

geneSetEnrich(
  x,
  celdaModel,
  useAssay = "counts",
  altExpName = "featureSubset",
  databases,
  fdr = 0.05
)

# S4 method for SingleCellExperiment
geneSetEnrich(
  x,
  useAssay = "counts",
  altExpName = "featureSubset",
  databases,
  fdr = 0.05
)

# S4 method for matrix
geneSetEnrich(x, celdaModel, databases, fdr = 0.05)

Arguments

x
A numeric matrix of counts or a SingleCellExperiment with the matrix located in the assay slot under useAssay. Rows represent features and columns represent cells. Rownames of the matrix or SingleCellExperiment object should be gene names.

celdaModel
Celda object of class celda_G or celda_CG.

useAssay
A string specifying which assay slot to use if x is a SingleCellExperiment object. Default "counts".

altExpName
The name for the altExp slot to use. Default "featureSubset".

databases
Character vector. Name of reference database. Available databases can be viewed by listEnrichrDbs.

fdr
False discovery rate (FDR). Numeric. Cutoff value for adjusted p-value, terms with FDR below this value are considered significantly enriched.

Value

List of length 'L' where each member contains the significantly enriched terms for the corresponding module.

Author

Ahmed Youssef, Zhe Wang

Examples

library(M3DExampleData)
counts <- M3DExampleData::Mmus_example_list$data
# subset 500 genes for fast clustering
counts <- counts[seq(1501, 2000), ]
# cluster genes into 10 modules for quick demo
sce <- celda_G(x = as.matrix(counts), L = 10, verbose = FALSE)
gse <- geneSetEnrich(sce,
  databases = c("GO_Biological_Process_2018", "GO_Molecular_Function_2018"))

#> EnrichR website not responding

#> EnrichR website not responding

#> EnrichR website not responding

#> EnrichR website not responding

#> EnrichR website not responding

#> EnrichR website not responding

#> EnrichR website not responding

#> EnrichR website not responding

#> EnrichR website not responding

#> EnrichR website not responding