Get feature, cell and sample names from a celdaModel

Retrieves the row, column, and sample names used to generate a celdaModel.

matrixNames(celdaMod)

# S4 method for celdaModel
matrixNames(celdaMod)

Arguments

celdaMod
celdaModel. Options available in `celda::availableModels`.

Value

List. Contains row, column, and sample character vectors corresponding to the values provided when the celdaModel was generated.

Examples

data(celdaCGMod)
matrixNames(celdaCGMod)

#> $row
#>   [1] "Gene_1"   "Gene_2"   "Gene_3"   "Gene_4"   "Gene_5"   "Gene_6"  
#>   [7] "Gene_7"   "Gene_8"   "Gene_9"   "Gene_10"  "Gene_11"  "Gene_12" 
#>  [13] "Gene_13"  "Gene_14"  "Gene_15"  "Gene_16"  "Gene_17"  "Gene_18" 
#>  [19] "Gene_19"  "Gene_20"  "Gene_21"  "Gene_22"  "Gene_23"  "Gene_24" 
#>  [25] "Gene_25"  "Gene_26"  "Gene_27"  "Gene_28"  "Gene_29"  "Gene_30" 
#>  [31] "Gene_31"  "Gene_32"  "Gene_33"  "Gene_34"  "Gene_35"  "Gene_36" 
#>  [37] "Gene_37"  "Gene_38"  "Gene_39"  "Gene_40"  "Gene_41"  "Gene_42" 
#>  [43] "Gene_43"  "Gene_44"  "Gene_45"  "Gene_46"  "Gene_47"  "Gene_48" 
#>  [49] "Gene_49"  "Gene_50"  "Gene_51"  "Gene_52"  "Gene_53"  "Gene_54" 
#>  [55] "Gene_55"  "Gene_56"  "Gene_57"  "Gene_58"  "Gene_59"  "Gene_60" 
#>  [61] "Gene_61"  "Gene_62"  "Gene_63"  "Gene_64"  "Gene_65"  "Gene_66" 
#>  [67] "Gene_67"  "Gene_68"  "Gene_69"  "Gene_70"  "Gene_71"  "Gene_72" 
#>  [73] "Gene_73"  "Gene_74"  "Gene_75"  "Gene_76"  "Gene_77"  "Gene_78" 
#>  [79] "Gene_79"  "Gene_80"  "Gene_81"  "Gene_82"  "Gene_83"  "Gene_84" 
#>  [85] "Gene_85"  "Gene_86"  "Gene_87"  "Gene_88"  "Gene_89"  "Gene_90" 
#>  [91] "Gene_91"  "Gene_92"  "Gene_93"  "Gene_94"  "Gene_95"  "Gene_96" 
#>  [97] "Gene_97"  "Gene_98"  "Gene_99"  "Gene_100"
#> 
#> $column
#>   [1] "Cell_1"   "Cell_2"   "Cell_3"   "Cell_4"   "Cell_5"   "Cell_6"  
#>   [7] "Cell_7"   "Cell_8"   "Cell_9"   "Cell_10"  "Cell_11"  "Cell_12" 
#>  [13] "Cell_13"  "Cell_14"  "Cell_15"  "Cell_16"  "Cell_17"  "Cell_18" 
#>  [19] "Cell_19"  "Cell_20"  "Cell_21"  "Cell_22"  "Cell_23"  "Cell_24" 
#>  [25] "Cell_25"  "Cell_26"  "Cell_27"  "Cell_28"  "Cell_29"  "Cell_30" 
#>  [31] "Cell_31"  "Cell_32"  "Cell_33"  "Cell_34"  "Cell_35"  "Cell_36" 
#>  [37] "Cell_37"  "Cell_38"  "Cell_39"  "Cell_40"  "Cell_41"  "Cell_42" 
#>  [43] "Cell_43"  "Cell_44"  "Cell_45"  "Cell_46"  "Cell_47"  "Cell_48" 
#>  [49] "Cell_49"  "Cell_50"  "Cell_51"  "Cell_52"  "Cell_53"  "Cell_54" 
#>  [55] "Cell_55"  "Cell_56"  "Cell_57"  "Cell_58"  "Cell_59"  "Cell_60" 
#>  [61] "Cell_61"  "Cell_62"  "Cell_63"  "Cell_64"  "Cell_65"  "Cell_66" 
#>  [67] "Cell_67"  "Cell_68"  "Cell_69"  "Cell_70"  "Cell_71"  "Cell_72" 
#>  [73] "Cell_73"  "Cell_74"  "Cell_75"  "Cell_76"  "Cell_77"  "Cell_78" 
#>  [79] "Cell_79"  "Cell_80"  "Cell_81"  "Cell_82"  "Cell_83"  "Cell_84" 
#>  [85] "Cell_85"  "Cell_86"  "Cell_87"  "Cell_88"  "Cell_89"  "Cell_90" 
#>  [91] "Cell_91"  "Cell_92"  "Cell_93"  "Cell_94"  "Cell_95"  "Cell_96" 
#>  [97] "Cell_97"  "Cell_98"  "Cell_99"  "Cell_100" "Cell_101" "Cell_102"
#> [103] "Cell_103" "Cell_104" "Cell_105" "Cell_106" "Cell_107" "Cell_108"
#> [109] "Cell_109" "Cell_110" "Cell_111" "Cell_112" "Cell_113" "Cell_114"
#> [115] "Cell_115" "Cell_116" "Cell_117" "Cell_118" "Cell_119" "Cell_120"
#> [121] "Cell_121" "Cell_122" "Cell_123" "Cell_124" "Cell_125" "Cell_126"
#> [127] "Cell_127" "Cell_128" "Cell_129" "Cell_130" "Cell_131" "Cell_132"
#> [133] "Cell_133" "Cell_134" "Cell_135" "Cell_136" "Cell_137" "Cell_138"
#> [139] "Cell_139" "Cell_140" "Cell_141" "Cell_142" "Cell_143" "Cell_144"
#> [145] "Cell_145" "Cell_146" "Cell_147" "Cell_148" "Cell_149" "Cell_150"
#> [151] "Cell_151" "Cell_152" "Cell_153" "Cell_154" "Cell_155" "Cell_156"
#> [157] "Cell_157" "Cell_158" "Cell_159" "Cell_160" "Cell_161" "Cell_162"
#> [163] "Cell_163" "Cell_164" "Cell_165" "Cell_166" "Cell_167" "Cell_168"
#> [169] "Cell_169" "Cell_170" "Cell_171" "Cell_172" "Cell_173" "Cell_174"
#> [175] "Cell_175" "Cell_176" "Cell_177" "Cell_178" "Cell_179" "Cell_180"
#> [181] "Cell_181" "Cell_182" "Cell_183" "Cell_184" "Cell_185" "Cell_186"
#> [187] "Cell_187" "Cell_188" "Cell_189" "Cell_190" "Cell_191" "Cell_192"
#> [193] "Cell_193" "Cell_194" "Cell_195" "Cell_196" "Cell_197" "Cell_198"
#> [199] "Cell_199" "Cell_200" "Cell_201" "Cell_202" "Cell_203" "Cell_204"
#> [205] "Cell_205" "Cell_206" "Cell_207" "Cell_208" "Cell_209" "Cell_210"
#> [211] "Cell_211" "Cell_212" "Cell_213" "Cell_214" "Cell_215" "Cell_216"
#> [217] "Cell_217" "Cell_218" "Cell_219" "Cell_220" "Cell_221" "Cell_222"
#> [223] "Cell_223" "Cell_224" "Cell_225" "Cell_226" "Cell_227" "Cell_228"
#> [229] "Cell_229" "Cell_230" "Cell_231" "Cell_232" "Cell_233" "Cell_234"
#> [235] "Cell_235" "Cell_236" "Cell_237" "Cell_238" "Cell_239" "Cell_240"
#> [241] "Cell_241" "Cell_242" "Cell_243" "Cell_244" "Cell_245" "Cell_246"
#> [247] "Cell_247" "Cell_248" "Cell_249" "Cell_250" "Cell_251" "Cell_252"
#> [253] "Cell_253" "Cell_254" "Cell_255" "Cell_256" "Cell_257" "Cell_258"
#> [259] "Cell_259" "Cell_260" "Cell_261" "Cell_262" "Cell_263" "Cell_264"
#> [265] "Cell_265" "Cell_266" "Cell_267" "Cell_268" "Cell_269" "Cell_270"
#> [271] "Cell_271" "Cell_272" "Cell_273" "Cell_274" "Cell_275" "Cell_276"
#> [277] "Cell_277" "Cell_278" "Cell_279" "Cell_280" "Cell_281" "Cell_282"
#> [283] "Cell_283" "Cell_284" "Cell_285" "Cell_286" "Cell_287" "Cell_288"
#> [289] "Cell_289" "Cell_290" "Cell_291" "Cell_292" "Cell_293" "Cell_294"
#> [295] "Cell_295" "Cell_296" "Cell_297" "Cell_298" "Cell_299" "Cell_300"
#> [301] "Cell_301" "Cell_302" "Cell_303" "Cell_304" "Cell_305" "Cell_306"
#> [307] "Cell_307" "Cell_308" "Cell_309" "Cell_310" "Cell_311" "Cell_312"
#> [313] "Cell_313" "Cell_314" "Cell_315" "Cell_316" "Cell_317" "Cell_318"
#> [319] "Cell_319" "Cell_320" "Cell_321" "Cell_322" "Cell_323" "Cell_324"
#> [325] "Cell_325" "Cell_326" "Cell_327" "Cell_328" "Cell_329" "Cell_330"
#> [331] "Cell_331" "Cell_332" "Cell_333" "Cell_334" "Cell_335" "Cell_336"
#> [337] "Cell_337" "Cell_338" "Cell_339" "Cell_340" "Cell_341" "Cell_342"
#> [343] "Cell_343" "Cell_344" "Cell_345" "Cell_346" "Cell_347" "Cell_348"
#> [349] "Cell_349" "Cell_350" "Cell_351" "Cell_352" "Cell_353" "Cell_354"
#> [355] "Cell_355" "Cell_356" "Cell_357" "Cell_358" "Cell_359" "Cell_360"
#> [361] "Cell_361" "Cell_362" "Cell_363" "Cell_364" "Cell_365" "Cell_366"
#> [367] "Cell_367" "Cell_368" "Cell_369" "Cell_370" "Cell_371" "Cell_372"
#> [373] "Cell_373" "Cell_374" "Cell_375" "Cell_376" "Cell_377" "Cell_378"
#> [379] "Cell_379" "Cell_380" "Cell_381" "Cell_382" "Cell_383" "Cell_384"
#> [385] "Cell_385" "Cell_386" "Cell_387" "Cell_388" "Cell_389" "Cell_390"
#> [391] "Cell_391" "Cell_392" "Cell_393" "Cell_394" "Cell_395" "Cell_396"
#> [397] "Cell_397" "Cell_398" "Cell_399" "Cell_400" "Cell_401" "Cell_402"
#> [403] "Cell_403" "Cell_404" "Cell_405" "Cell_406" "Cell_407" "Cell_408"
#> [409] "Cell_409" "Cell_410" "Cell_411" "Cell_412" "Cell_413" "Cell_414"
#> [415] "Cell_415" "Cell_416" "Cell_417" "Cell_418" "Cell_419" "Cell_420"
#> [421] "Cell_421" "Cell_422" "Cell_423" "Cell_424" "Cell_425"
#> 
#> $sample
#> [1] "Sample_1" "Sample_2" "Sample_3" "Sample_4" "Sample_5"
#>