Get run parameters from a celda model SingleCellExperiment or celdaList object

Returns details on the clustering parameters and model priors from the celdaList object when it was created.

runParams(x, altExpName = "featureSubset")

# S4 method for SingleCellExperiment
runParams(x, altExpName = "featureSubset")

# S4 method for celdaList
runParams(x)

Arguments

x

An object of class SingleCellExperiment or class celdaList.

altExpName

The name for the altExp slot to use. Default "featureSubset".

Value

Data Frame. Contains details on the various K/L parameters, chain parameters, seed, and final log-likelihoods derived for each model in the provided celdaList.

Examples

data(sceCeldaCGGridSearch)
runParams(sceCeldaCGGridSearch)
#>   index chain K  L  seed logLikelihood mean_perplexity
#> 1     1     1 4  9 12345      -1226407        46.95862
#> 2     2     1 5  9 12345      -1214714        45.25311
#> 3     3     1 6  9 12345      -1214754        45.25124
#> 4     4     1 4 10 12345      -1225233        46.77100
#> 5     5     1 5 10 12345      -1212891        44.98559
#> 6     6     1 6 10 12345      -1212928        44.98491
#> 7     7     1 4 11 12345      -1225255        46.77053
#> 8     8     1 5 11 12345      -1212917        44.98514
#> 9     9     1 6 11 12345      -1212953        44.98495
data(celdaCGGridSearchRes)
runParams(celdaCGGridSearchRes)
#>   index chain K  L logLikelihood mean_perplexity
#> 1     1     1 4  9      -1228381        47.12902
#> 2     2     1 5  9      -1217364        45.51741
#> 3     3     1 6  9      -1217407        45.51451
#> 4     4     1 4 10      -1227891        47.04897
#> 5     5     1 5 10      -1215541        45.24781
#> 6     6     1 6 10      -1215583        45.24439
#> 7     7     1 4 11      -1227913        47.04849
#> 8     8     1 5 11      -1215567        45.24733
#> 9     9     1 6 11      -1215619        45.24380