A function to draw clustered heatmaps.

A function to draw clustered heatmaps where one has better control over some graphical parameters such as cell size, etc.

semi_pheatmap(mat, color = colorRampPalette(rev(brewer.pal(n = 7, name =
  "RdYlBu")))(100), kmeans_k = NA, breaks = NA, border_color = "grey60",
  cellwidth = NA, cellheight = NA, scale = "none", cluster_rows = TRUE,
  cluster_cols = TRUE, clustering_distance_rows = "euclidean",
  clustering_distance_cols = "euclidean", clustering_method = "complete",
  clustering_callback = identity2, cutree_rows = NA, cutree_cols = NA,
  treeheight_row = ifelse(cluster_rows, 50, 0),
  treeheight_col = ifelse(cluster_cols, 50, 0), legend = TRUE,
  legend_breaks = NA, legend_labels = NA, annotation_row = NA,
  annotation_col = NA, annotation = NA, annotation_colors = NA,
  annotation_legend = TRUE, annotation_names_row = TRUE,
  annotation_names_col = TRUE, drop_levels = TRUE, show_rownames = TRUE,
  show_colnames = TRUE, main = NA, fontsize = 10,
  fontsize_row = fontsize, fontsize_col = fontsize,
  display_numbers = FALSE, number_format = "%.2f",
  number_color = "grey30", fontsize_number = 0.8 * fontsize,
  gaps_row = NULL, gaps_col = NULL, labels_row = NULL,
  labels_col = NULL, filename = NA, width = NA, height = NA,
  silent = FALSE, row_label, col_label, ...)

Arguments

mat
numeric matrix of the values to be plotted.

color
vector of colors used in heatmap.

kmeans_k
the number of kmeans clusters to make, if we want to agggregate the rows before drawing heatmap. If NA then the rows are not aggregated.

breaks
Numeric vector. A sequence of numbers that covers the range of values in the normalized `counts`. Values in the normalized `matrix` are assigned to each bin in `breaks`. Each break is assigned to a unique color from `col`. If NULL, then breaks are calculated automatically. Default NULL.

border_color
color of cell borders on heatmap, use NA if no border should be drawn.

cellwidth
individual cell width in points. If left as NA, then the values depend on the size of plotting window.

cellheight
individual cell height in points. If left as NA, then the values depend on the size of plotting window.

scale
character indicating if the values should be centered and scaled in either the row direction or the column direction, or none. Corresponding values are "row", "column" and "none"

cluster_rows
boolean values determining if rows should be clustered or hclust object,

cluster_cols
boolean values determining if columns should be clustered or hclust object.

clustering_distance_rows
distance measure used in clustering rows. Possible values are "correlation" for Pearson correlation and all the distances supported by dist, such as "euclidean", etc. If the value is none of the above it is assumed that a distance matrix is provided.

clustering_distance_cols
distance measure used in clustering columns. Possible values the same as for clustering_distance_rows.

clustering_method
clustering method used. Accepts the same values as hclust.

clustering_callback
callback function to modify the clustering. Is called with two parameters: original hclust object and the matrix used for clustering. Must return a hclust object.

cutree_rows
number of clusters the rows are divided into, based on the hierarchical clustering (using cutree), if rows are not clustered, the argument is ignored

cutree_cols
similar to cutree_rows, but for columns

treeheight_row
the height of a tree for rows, if these are clustered. Default value 50 points.

treeheight_col
the height of a tree for columns, if these are clustered. Default value 50 points.

legend
logical to determine if legend should be drawn or not.

legend_breaks
vector of breakpoints for the legend.

legend_labels
vector of labels for the legend_breaks.

annotation_row
data frame that specifies the annotations shown on left side of the heatmap. Each row defines the features for a specific row. The rows in the data and in the annotation are matched using corresponding row names. Note that color schemes takes into account if variable is continuous or discrete.

annotation_col
similar to annotation_row, but for columns.

annotation
deprecated parameter that currently sets the annotation_col if it is missing

annotation_colors
list for specifying annotation_row and annotation_col track colors manually. It is possible to define the colors for only some of the features. Check examples for details.

annotation_legend
boolean value showing if the legend for annotation tracks should be drawn.

annotation_names_row
boolean value showing if the names for row annotation tracks should be drawn.

annotation_names_col
boolean value showing if the names for column annotation tracks should be drawn.

drop_levels
logical to determine if unused levels are also shown in the legend

show_rownames
boolean specifying if column names are be shown.

show_colnames
boolean specifying if column names are be shown.

main
the title of the plot

fontsize
base fontsize for the plot

fontsize_row
fontsize for rownames (Default: fontsize)

fontsize_col
fontsize for colnames (Default: fontsize)

display_numbers
logical determining if the numeric values are also printed to the cells. If this is a matrix (with same dimensions as original matrix), the contents of the matrix are shown instead of original values.

number_format
format strings (C printf style) of the numbers shown in cells. For example "%.2f" shows 2 decimal places and "%.1e" shows exponential notation (see more in sprintf).

number_color
color of the text

fontsize_number
fontsize of the numbers displayed in cells

gaps_row
vector of row indices that show shere to put gaps into heatmap. Used only if the rows are not clustered. See cutree_row to see how to introduce gaps to clustered rows.

gaps_col
similar to gaps_row, but for columns.

labels_row
custom labels for rows that are used instead of rownames.

labels_col
similar to labels_row, but for columns.

filename
file path where to save the picture. Filetype is decided by the extension in the path. Currently following formats are supported: png, pdf, tiff, bmp, jpeg. Even if the plot does not fit into the plotting window, the file size is calculated so that the plot would fit there, unless specified otherwise.

width
manual option for determining the output file width in inches.

height
manual option for determining the output file height in inches.

silent
do not draw the plot (useful when using the gtable output)

row_label
row cluster labels for semi-clustering

col_label
column cluster labels for semi-clustering

…
graphical parameters for the text used in plot. Parameters passed to grid.text, see gpar.

Value

Invisibly a list of components

tree_row the clustering of rows as hclust object

tree_col the clustering of columns as hclust object

kmeans the kmeans clustering of rows if parameter kmeans_k was specified

Details

The function also allows to aggregate the rows using kmeans clustering. This is advisable if number of rows is so big that R cannot handle their hierarchical clustering anymore, roughly more than 1000. Instead of showing all the rows separately one can cluster the rows in advance and show only the cluster centers. The number of clusters can be tuned with parameter kmeans_k.