Output a feature module table

Creates a table that contains the list of features in each feature module.

featureModuleTable(
  sce,
  useAssay = "counts",
  altExpName = "featureSubset",
  displayName = NULL,
  outputFile = NULL
)

Arguments

sce

A SingleCellExperiment object returned by celda_G, or celda_CG, with the matrix located in the useAssay assay slot. Rows represent features and columns represent cells.

useAssay

A string specifying which assay slot to use. Default "counts".

altExpName

The name for the altExp slot to use. Default "featureSubset".

displayName

Character. The column name of rowData(sce) that specifies the display names for the features. Default NULL, which displays the row names.

outputFile

File name for feature module table. If NULL, file will not be created. Default NULL.

Value

Matrix. Contains a list of features per each column (feature module)

Examples

data(sceCeldaCG)
featureModuleTable(sceCeldaCG)
#>       L1        L2        L3        L4         L5        L6        L7       
#>  [1,] "Gene_45" "Gene_77" "Gene_72" "Gene_87"  "Gene_41" "Gene_80" "Gene_24"
#>  [2,] "Gene_5"  "Gene_17" "Gene_15" "Gene_6"   "Gene_70" "Gene_60" "Gene_53"
#>  [3,] "Gene_94" "Gene_30" "Gene_22" "Gene_82"  "Gene_64" "Gene_55" "Gene_74"
#>  [4,] "Gene_91" "Gene_9"  "Gene_52" "Gene_28"  "Gene_36" "Gene_40" "Gene_59"
#>  [5,] "Gene_3"  ""        "Gene_78" "Gene_39"  "Gene_37" "Gene_25" "Gene_89"
#>  [6,] "Gene_50" ""        "Gene_84" "Gene_10"  "Gene_14" "Gene_62" "Gene_11"
#>  [7,] "Gene_90" ""        "Gene_81" "Gene_51"  "Gene_99" "Gene_97" "Gene_49"
#>  [8,] ""        ""        ""        "Gene_100" "Gene_20" "Gene_95" "Gene_1" 
#>  [9,] ""        ""        ""        "Gene_65"  "Gene_2"  "Gene_46" "Gene_79"
#> [10,] ""        ""        ""        "Gene_33"  "Gene_44" "Gene_76" ""       
#> [11,] ""        ""        ""        "Gene_38"  "Gene_66" ""        ""       
#> [12,] ""        ""        ""        ""         "Gene_32" ""        ""       
#> [13,] ""        ""        ""        ""         "Gene_34" ""        ""       
#> [14,] ""        ""        ""        ""         "Gene_23" ""        ""       
#> [15,] ""        ""        ""        ""         ""        ""        ""       
#> [16,] ""        ""        ""        ""         ""        ""        ""       
#>       L8        L9        L10      
#>  [1,] "Gene_4"  "Gene_85" "Gene_69"
#>  [2,] "Gene_43" "Gene_98" "Gene_47"
#>  [3,] "Gene_42" "Gene_35" "Gene_13"
#>  [4,] "Gene_29" "Gene_8"  "Gene_92"
#>  [5,] "Gene_73" "Gene_83" "Gene_88"
#>  [6,] "Gene_61" "Gene_19" "Gene_93"
#>  [7,] ""        "Gene_63" "Gene_12"
#>  [8,] ""        "Gene_75" "Gene_86"
#>  [9,] ""        "Gene_21" "Gene_58"
#> [10,] ""        "Gene_48" "Gene_18"
#> [11,] ""        "Gene_31" "Gene_7" 
#> [12,] ""        "Gene_57" "Gene_16"
#> [13,] ""        "Gene_27" ""       
#> [14,] ""        "Gene_67" ""       
#> [15,] ""        "Gene_68" ""       
#> [16,] ""        "Gene_54" ""