Get or set sample labels from a celda SingleCellExperiment object

Return or set the sample labels for the cells in sce.

sampleLabel(x, altExpName = "featureSubset")

# S4 method for SingleCellExperiment
sampleLabel(x, altExpName = "featureSubset")

sampleLabel(x, altExpName = "featureSubset") <- value

# S4 method for SingleCellExperiment
sampleLabel(x, altExpName = "featureSubset") <- value

# S4 method for celdaModel
sampleLabel(x)

Arguments

x

Can be one of

A SingleCellExperiment object returned by celda_C, celda_G, or celda_CG, with the matrix located in the useAssay assay slot. Rows represent features and columns represent cells.

A celda model object.

altExpName

The name for the altExp slot to use. Default "featureSubset".

value

Character vector of sample labels for replacements. Works only is x is a SingleCellExperiment object.

Value

Character vector. Contains the sample labels provided at model creation, or those automatically generated by celda.

Examples

data(sceCeldaCG)
sampleLabel(sceCeldaCG)
#>   [1] Sample_1 Sample_1 Sample_1 Sample_1 Sample_1 Sample_1 Sample_1 Sample_1
#>   [9] Sample_1 Sample_1 Sample_1 Sample_1 Sample_1 Sample_1 Sample_1 Sample_1
#>  [17] Sample_1 Sample_1 Sample_1 Sample_1 Sample_1 Sample_1 Sample_1 Sample_1
#>  [25] Sample_1 Sample_1 Sample_1 Sample_1 Sample_1 Sample_1 Sample_1 Sample_1
#>  [33] Sample_1 Sample_1 Sample_1 Sample_1 Sample_1 Sample_1 Sample_1 Sample_1
#>  [41] Sample_1 Sample_1 Sample_1 Sample_1 Sample_1 Sample_1 Sample_1 Sample_1
#>  [49] Sample_1 Sample_1 Sample_1 Sample_1 Sample_1 Sample_1 Sample_1 Sample_1
#>  [57] Sample_1 Sample_1 Sample_1 Sample_1 Sample_1 Sample_1 Sample_1 Sample_1
#>  [65] Sample_1 Sample_1 Sample_1 Sample_1 Sample_1 Sample_1 Sample_1 Sample_1
#>  [73] Sample_1 Sample_1 Sample_1 Sample_1 Sample_1 Sample_1 Sample_1 Sample_1
#>  [81] Sample_1 Sample_1 Sample_1 Sample_1 Sample_1 Sample_1 Sample_1 Sample_1
#>  [89] Sample_1 Sample_1 Sample_1 Sample_1 Sample_1 Sample_2 Sample_2 Sample_2
#>  [97] Sample_2 Sample_2 Sample_2 Sample_2 Sample_2 Sample_2 Sample_2 Sample_2
#> [105] Sample_2 Sample_2 Sample_2 Sample_2 Sample_2 Sample_2 Sample_2 Sample_2
#> [113] Sample_2 Sample_2 Sample_2 Sample_2 Sample_2 Sample_2 Sample_2 Sample_2
#> [121] Sample_2 Sample_2 Sample_2 Sample_2 Sample_2 Sample_2 Sample_2 Sample_2
#> [129] Sample_2 Sample_2 Sample_2 Sample_2 Sample_2 Sample_2 Sample_2 Sample_2
#> [137] Sample_2 Sample_2 Sample_2 Sample_2 Sample_2 Sample_2 Sample_2 Sample_2
#> [145] Sample_2 Sample_2 Sample_2 Sample_2 Sample_2 Sample_2 Sample_2 Sample_2
#> [153] Sample_2 Sample_2 Sample_2 Sample_2 Sample_2 Sample_2 Sample_2 Sample_2
#> [161] Sample_2 Sample_2 Sample_2 Sample_2 Sample_2 Sample_2 Sample_2 Sample_2
#> [169] Sample_2 Sample_2 Sample_2 Sample_2 Sample_2 Sample_2 Sample_2 Sample_2
#> [177] Sample_2 Sample_2 Sample_2 Sample_2 Sample_2 Sample_2 Sample_2 Sample_2
#> [185] Sample_2 Sample_2 Sample_2 Sample_2 Sample_3 Sample_3 Sample_3 Sample_3
#> [193] Sample_3 Sample_3 Sample_3 Sample_3 Sample_3 Sample_3 Sample_3 Sample_3
#> [201] Sample_3 Sample_3 Sample_3 Sample_3 Sample_3 Sample_3 Sample_3 Sample_3
#> [209] Sample_3 Sample_3 Sample_3 Sample_3 Sample_3 Sample_3 Sample_3 Sample_3
#> [217] Sample_3 Sample_3 Sample_3 Sample_3 Sample_3 Sample_3 Sample_3 Sample_3
#> [225] Sample_3 Sample_3 Sample_3 Sample_3 Sample_3 Sample_3 Sample_3 Sample_3
#> [233] Sample_3 Sample_3 Sample_3 Sample_3 Sample_3 Sample_3 Sample_3 Sample_3
#> [241] Sample_3 Sample_3 Sample_3 Sample_3 Sample_3 Sample_3 Sample_3 Sample_3
#> [249] Sample_3 Sample_3 Sample_3 Sample_3 Sample_3 Sample_3 Sample_3 Sample_3
#> [257] Sample_3 Sample_3 Sample_3 Sample_3 Sample_3 Sample_3 Sample_3 Sample_3
#> [265] Sample_3 Sample_3 Sample_3 Sample_3 Sample_3 Sample_3 Sample_3 Sample_3
#> [273] Sample_3 Sample_3 Sample_3 Sample_3 Sample_3 Sample_3 Sample_3 Sample_3
#> [281] Sample_3 Sample_3 Sample_4 Sample_4 Sample_4 Sample_4 Sample_4 Sample_4
#> [289] Sample_4 Sample_4 Sample_4 Sample_4 Sample_4 Sample_4 Sample_4 Sample_4
#> [297] Sample_4 Sample_4 Sample_4 Sample_4 Sample_4 Sample_4 Sample_4 Sample_4
#> [305] Sample_4 Sample_4 Sample_4 Sample_4 Sample_4 Sample_4 Sample_4 Sample_4
#> [313] Sample_4 Sample_4 Sample_4 Sample_4 Sample_4 Sample_4 Sample_4 Sample_4
#> [321] Sample_4 Sample_4 Sample_4 Sample_4 Sample_4 Sample_4 Sample_4 Sample_4
#> [329] Sample_4 Sample_4 Sample_4 Sample_4 Sample_4 Sample_4 Sample_4 Sample_4
#> [337] Sample_4 Sample_4 Sample_4 Sample_4 Sample_4 Sample_4 Sample_4 Sample_4
#> [345] Sample_4 Sample_4 Sample_4 Sample_4 Sample_4 Sample_5 Sample_5 Sample_5
#> [353] Sample_5 Sample_5 Sample_5 Sample_5 Sample_5 Sample_5 Sample_5 Sample_5
#> [361] Sample_5 Sample_5 Sample_5 Sample_5 Sample_5 Sample_5 Sample_5 Sample_5
#> [369] Sample_5 Sample_5 Sample_5 Sample_5 Sample_5 Sample_5 Sample_5 Sample_5
#> [377] Sample_5 Sample_5 Sample_5 Sample_5 Sample_5 Sample_5 Sample_5 Sample_5
#> [385] Sample_5 Sample_5 Sample_5 Sample_5 Sample_5 Sample_5 Sample_5 Sample_5
#> [393] Sample_5 Sample_5 Sample_5 Sample_5 Sample_5 Sample_5 Sample_5 Sample_5
#> [401] Sample_5 Sample_5 Sample_5 Sample_5 Sample_5 Sample_5 Sample_5 Sample_5
#> [409] Sample_5 Sample_5 Sample_5 Sample_5 Sample_5 Sample_5 Sample_5 Sample_5
#> [417] Sample_5 Sample_5 Sample_5 Sample_5 Sample_5 Sample_5 Sample_5 Sample_5
#> [425] Sample_5
#> Levels: Sample_1 Sample_2 Sample_3 Sample_4 Sample_5
data(celdaCGMod)
sampleLabel(celdaCGMod)
#>   [1] Sample_1 Sample_1 Sample_1 Sample_1 Sample_1 Sample_1 Sample_1 Sample_1
#>   [9] Sample_1 Sample_1 Sample_1 Sample_1 Sample_1 Sample_1 Sample_1 Sample_1
#>  [17] Sample_1 Sample_1 Sample_1 Sample_1 Sample_1 Sample_1 Sample_1 Sample_1
#>  [25] Sample_1 Sample_1 Sample_1 Sample_1 Sample_1 Sample_1 Sample_1 Sample_1
#>  [33] Sample_1 Sample_1 Sample_1 Sample_1 Sample_1 Sample_1 Sample_1 Sample_1
#>  [41] Sample_1 Sample_1 Sample_1 Sample_1 Sample_1 Sample_1 Sample_1 Sample_1
#>  [49] Sample_1 Sample_1 Sample_1 Sample_1 Sample_1 Sample_1 Sample_1 Sample_1
#>  [57] Sample_1 Sample_1 Sample_1 Sample_1 Sample_1 Sample_1 Sample_1 Sample_1
#>  [65] Sample_1 Sample_1 Sample_1 Sample_1 Sample_1 Sample_1 Sample_1 Sample_1
#>  [73] Sample_1 Sample_1 Sample_1 Sample_1 Sample_1 Sample_1 Sample_1 Sample_1
#>  [81] Sample_1 Sample_1 Sample_1 Sample_1 Sample_1 Sample_1 Sample_1 Sample_1
#>  [89] Sample_1 Sample_1 Sample_1 Sample_1 Sample_1 Sample_2 Sample_2 Sample_2
#>  [97] Sample_2 Sample_2 Sample_2 Sample_2 Sample_2 Sample_2 Sample_2 Sample_2
#> [105] Sample_2 Sample_2 Sample_2 Sample_2 Sample_2 Sample_2 Sample_2 Sample_2
#> [113] Sample_2 Sample_2 Sample_2 Sample_2 Sample_2 Sample_2 Sample_2 Sample_2
#> [121] Sample_2 Sample_2 Sample_2 Sample_2 Sample_2 Sample_2 Sample_2 Sample_2
#> [129] Sample_2 Sample_2 Sample_2 Sample_2 Sample_2 Sample_2 Sample_2 Sample_2
#> [137] Sample_2 Sample_2 Sample_2 Sample_2 Sample_2 Sample_2 Sample_2 Sample_2
#> [145] Sample_2 Sample_2 Sample_2 Sample_2 Sample_2 Sample_2 Sample_2 Sample_2
#> [153] Sample_2 Sample_2 Sample_2 Sample_2 Sample_2 Sample_2 Sample_2 Sample_2
#> [161] Sample_2 Sample_2 Sample_2 Sample_2 Sample_2 Sample_2 Sample_2 Sample_2
#> [169] Sample_2 Sample_2 Sample_2 Sample_2 Sample_2 Sample_2 Sample_2 Sample_2
#> [177] Sample_2 Sample_2 Sample_2 Sample_2 Sample_2 Sample_2 Sample_2 Sample_2
#> [185] Sample_2 Sample_2 Sample_2 Sample_2 Sample_3 Sample_3 Sample_3 Sample_3
#> [193] Sample_3 Sample_3 Sample_3 Sample_3 Sample_3 Sample_3 Sample_3 Sample_3
#> [201] Sample_3 Sample_3 Sample_3 Sample_3 Sample_3 Sample_3 Sample_3 Sample_3
#> [209] Sample_3 Sample_3 Sample_3 Sample_3 Sample_3 Sample_3 Sample_3 Sample_3
#> [217] Sample_3 Sample_3 Sample_3 Sample_3 Sample_3 Sample_3 Sample_3 Sample_3
#> [225] Sample_3 Sample_3 Sample_3 Sample_3 Sample_3 Sample_3 Sample_3 Sample_3
#> [233] Sample_3 Sample_3 Sample_3 Sample_3 Sample_3 Sample_3 Sample_3 Sample_3
#> [241] Sample_3 Sample_3 Sample_3 Sample_3 Sample_3 Sample_3 Sample_3 Sample_3
#> [249] Sample_3 Sample_3 Sample_3 Sample_3 Sample_3 Sample_3 Sample_3 Sample_3
#> [257] Sample_3 Sample_3 Sample_3 Sample_3 Sample_3 Sample_3 Sample_3 Sample_3
#> [265] Sample_3 Sample_3 Sample_3 Sample_3 Sample_3 Sample_3 Sample_3 Sample_3
#> [273] Sample_3 Sample_3 Sample_3 Sample_3 Sample_3 Sample_3 Sample_3 Sample_3
#> [281] Sample_3 Sample_3 Sample_4 Sample_4 Sample_4 Sample_4 Sample_4 Sample_4
#> [289] Sample_4 Sample_4 Sample_4 Sample_4 Sample_4 Sample_4 Sample_4 Sample_4
#> [297] Sample_4 Sample_4 Sample_4 Sample_4 Sample_4 Sample_4 Sample_4 Sample_4
#> [305] Sample_4 Sample_4 Sample_4 Sample_4 Sample_4 Sample_4 Sample_4 Sample_4
#> [313] Sample_4 Sample_4 Sample_4 Sample_4 Sample_4 Sample_4 Sample_4 Sample_4
#> [321] Sample_4 Sample_4 Sample_4 Sample_4 Sample_4 Sample_4 Sample_4 Sample_4
#> [329] Sample_4 Sample_4 Sample_4 Sample_4 Sample_4 Sample_4 Sample_4 Sample_4
#> [337] Sample_4 Sample_4 Sample_4 Sample_4 Sample_4 Sample_4 Sample_4 Sample_4
#> [345] Sample_4 Sample_4 Sample_4 Sample_4 Sample_4 Sample_5 Sample_5 Sample_5
#> [353] Sample_5 Sample_5 Sample_5 Sample_5 Sample_5 Sample_5 Sample_5 Sample_5
#> [361] Sample_5 Sample_5 Sample_5 Sample_5 Sample_5 Sample_5 Sample_5 Sample_5
#> [369] Sample_5 Sample_5 Sample_5 Sample_5 Sample_5 Sample_5 Sample_5 Sample_5
#> [377] Sample_5 Sample_5 Sample_5 Sample_5 Sample_5 Sample_5 Sample_5 Sample_5
#> [385] Sample_5 Sample_5 Sample_5 Sample_5 Sample_5 Sample_5 Sample_5 Sample_5
#> [393] Sample_5 Sample_5 Sample_5 Sample_5 Sample_5 Sample_5 Sample_5 Sample_5
#> [401] Sample_5 Sample_5 Sample_5 Sample_5 Sample_5 Sample_5 Sample_5 Sample_5
#> [409] Sample_5 Sample_5 Sample_5 Sample_5 Sample_5 Sample_5 Sample_5 Sample_5
#> [417] Sample_5 Sample_5 Sample_5 Sample_5 Sample_5 Sample_5 Sample_5 Sample_5
#> [425] Sample_5
#> Levels: Sample_1 Sample_2 Sample_3 Sample_4 Sample_5